Added Loading component
During fetch data from API, the Loading component is showing and when data came in, results are showing.
When fetching is started, the initial value of the isLoading state changing from false to true, and after assigning values from fetched data to results, the state of isLoading changes back to false.